requirements.txt
paddlepaddle-gpu==2.5.2
paddleocr>=2.7.0.3
pdf2image==1.16.3
Pillow==9.5.0
opencv-python==4.7.0.72
numpy
requests